Module tower_lsp::jsonrpc

A subset of JSON-RPC types used by the Language Server Protocol.

Structs

Error

A JSON-RPC error object.

Request

A JSON-RPC request or notification.

RequestBuilder

A builder to construct the properties of a Request.

Response

A successful or failed JSON-RPC response.

Enums

ErrorCode

A list of numeric error codes used in JSON-RPC responses.

Id

A unique ID used to correlate requests and responses together.

Traits

FromParams

A trait implemented by all JSON-RPC method parameters.

IntoResponse

A trait implemented by all JSON-RPC response types.

Method

A trait implemented by all valid JSON-RPC method handlers.

Type Aliases

Result

A specialized Result error type for JSON-RPC handlers.