Struct tower_lsp::jsonrpc::Response

pub struct Response { /* private fields */ }

A successful or failed JSON-RPC response.

Implementations

impl Response

pub const fn from_ok(id: Id, result: Value) -> Self

Creates a new successful response from a request ID and Error object.

pub const fn from_error(id: Id, error: Error) -> Self

Creates a new error response from a request ID and Error object.

pub fn from_parts(id: Id, body: Result<Value>) -> Self

Creates a new response from a request ID and either an Ok(Value) or Err(Error) body.

pub fn into_parts(self) -> (Id, Result<Value>)

Splits the response into a request ID paired with either an Ok(Value) or Err(Error) to signify whether the response is a success or failure.

pub const fn is_ok(&self) -> bool

Returns true if the response indicates success.

pub const fn is_error(&self) -> bool

Returns true if the response indicates failure.

pub const fn result(&self) -> Option<&Value>

Returns the result value, if it exists.

This member only exists if the response indicates success.

pub const fn error(&self) -> Option<&Error>

Returns the error value, if it exists.

This member only exists if the response indicates failure.

pub const fn id(&self) -> &Id

Returns the corresponding request ID, if known.