Trait tower_lsp::jsonrpc::Method

pub trait Method<S, P, R>: Sealed {
    type Future: Future<Output = R> + Send;

    // Required method
    fn invoke(&self, server: S, params: P) -> Self::Future;
}

A trait implemented by all valid JSON-RPC method handlers.

This trait abstracts over the following classes of functions and/or closures:

Signature	Description

async fn f(&self) -> jsonrpc::Result<R>	Request without parameters

async fn f(&self, params: P) -> jsonrpc::Result<R>	Request with required parameters

async fn f(&self)	Notification without parameters

async fn f(&self, params: P)	Notification with parameters

Required Associated Types

type Future: Future<Output = R> + Send

The future response value.

Required Methods

fn invoke(&self, server: S, params: P) -> Self::Future

Invokes the method with the given server receiver and parameters.