Struct tower_lsp::LspService

pub struct LspService<S> { /* private fields */ }

Service abstraction for the Language Server Protocol.

This service takes an incoming JSON-RPC message as input and produces an outgoing message as output. If the incoming message is a server notification or a client response, then the corresponding response will be None.

This implements tower::Service in order to remain independent from the underlying transport and to facilitate further abstraction with middleware.

Pending requests can be canceled by issuing a $/cancelRequest notification.

The service shuts down and stops serving requests after the exit notification is received.

Implementations

impl<S: LanguageServer> LspService<S>

pub fn new<F>(init: F) -> (Self, ClientSocket) where
F: FnOnce(Client) -> S,

Creates a new LspService with the given server backend, also returning a channel for server-to-client communication.

pub fn build<F>(init: F) -> LspServiceBuilder<S> where
F: FnOnce(Client) -> S,

Starts building a new LspService.

Returns an LspServiceBuilder, which allows adding custom JSON-RPC methods to the server.