Trait tower_lsp::jsonrpc::IntoResponse

pub trait IntoResponse: Sealed + Send + 'static {
    fn into_response(self, id: Option<Id>) -> Option<Response>;
    fn is_notification() -> bool;
}

A trait implemented by all JSON-RPC response types.

Required Methods

fn into_response(self, id: Option<Id>) -> Option<Response>

Attempts to construct a Response using Self and a corresponding Id.

fn is_notification() -> bool

Returns true if this is a notification response type.